+++ lib/features/offeredfood/data/repository/firebase_offered_food_repository.dart
@@ -178,6 +178,7 @@ class FirebaseOfferedFoodRepository implements OfferedFoodRepository {
           consumeBy: element.consumeBy?.getDate(),
           foodBoxId: boxId,
           foodBoxCount: boxCount,
+          foodTemperature: element.foodTemperature,
         ),
       );