+++ lib/ui/widgets/assist_chip.dart
@@ -26,19 +26,56 @@ class AssistChip extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final colors = _AssistChipColors.fromSelected(selected);
     return ActionChip(
-      label: Text(text),
+      label: Text(
+        text,
+        style: TextStyle(color: colors.textColor),
+      ),
       materialTapTargetSize: MaterialTapTargetSize.shrinkWrap,
       onPressed: onPressed,
-      backgroundColor:
-          selected ? ZOColors.assistChipSelectedBackground : Colors.transparent,
+      backgroundColor: colors.backgroundColor,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(8.0),
         side: BorderSide(
-          color: selected ? Colors.black : ZOColors.outline,
+          color: colors.borderColor,
           width: 1.0,
         ),
       ),
     );
+  }
+}
+
+/// A class that holds the colors for the [AssistChip] widget.
+class _AssistChipColors {
+  /// The text color of the chip.
+  final Color textColor;
+
+  /// The background color of the chip.
+  final Color backgroundColor;
+
+  /// The border color of the chip.
+  final Color borderColor;
+
+  /// Creates an instance of [_AssistChipColors].
+  const _AssistChipColors({
+    required this.textColor,
+    required this.backgroundColor,
+    required this.borderColor,
+  });
+
+  /// Creates an instance of [_AssistChipColors] based on the selected state.
+  factory _AssistChipColors.fromSelected(bool selected) {
+    return selected
+        ? const _AssistChipColors(
+            textColor: ZOColors.onSecondary,
+            backgroundColor: ZOColors.secondary,
+            borderColor: ZOColors.onPrimaryLight,
+          )
+        : const _AssistChipColors(
+            textColor: ZOColors.onPrimaryLight,
+            backgroundColor: Colors.transparent,
+            borderColor: ZOColors.outline,
+          );
   }
 }